+++ Samples~/Scripts/AssetPreviewExample.cs
@@ -4,7 +4,7 @@ namespace SaintsField.Samples.Scripts
 {
     public class AssetPreviewExample: MonoBehaviour
     {
-        [AssetPreview(50, align: EAlign.FieldStart)] public GameObject _go;
+        [AssetPreview(int.MaxValue, 500, align: EAlign.FieldStart)] public GameObject _go;
         [AssetPreview(20, 100)] public Texture2D drawTexture2D;
         [AssetPreview(above: true)] public Sprite _sprite;
         [AssetPreview(above: true)] public Sprite[] sprites;